Breakout algorithms into group files
